@@ -10,7 +10,7 @@ import { Editors } from '../editor/editors';
 import { Code } from '../../vscode/code';
 import { QuickOpen } from '../quickopen/quickopen';
 
-export enum ActivityBarPosition {
+export const enum ActivityBarPosition {
 	LEFT = 0,
 	RIGHT = 1
 }
@@ -42,6 +42,6 @@ export class SettingsEditor {
 	}
 
 	private async openSettings(): Promise<void> {
-		await this.quickopen.runCommand('Preferences: Open User Settings');
+		await this.quickopen.runCommand('Preferences: Open Settings (JSON)');
 	}
 }
