@@ -8,8 +8,8 @@ import { Code } from '../../vscode/code';
 
 export class QuickOpen {
 
+	static QUICK_OPEN = 'div.monaco-quick-open-widget';
 	static QUICK_OPEN_HIDDEN = 'div.monaco-quick-open-widget[aria-hidden="true"]';
-	static QUICK_OPEN = 'div.monaco-quick-open-widget[aria-hidden="false"]';
 	static QUICK_OPEN_INPUT = `${QuickOpen.QUICK_OPEN} .quick-open-input input`;
 	static QUICK_OPEN_FOCUSED_ELEMENT = `${QuickOpen.QUICK_OPEN} .quick-open-tree .monaco-tree-row.focused .monaco-highlighted-label`;
 	static QUICK_OPEN_ENTRY_SELECTOR = 'div[aria-label="Quick Picker"] .monaco-tree-rows.show-twisties .monaco-tree-row .quick-open-entry';
