@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Code } from '../../vscode/code';
+import { QuickOpen } from '../quickopen/quickopen';
 
 const PANEL_SELECTOR = 'div[id="workbench.panel.terminal"]';
 const XTERM_SELECTOR = `${PANEL_SELECTOR} .terminal-wrapper`;
@@ -11,10 +12,10 @@ const XTERM_TEXTAREA = `${XTERM_SELECTOR} textarea.xterm-helper-textarea`;
 
 export class Terminal {
 
-	constructor(private code: Code) { }
+	constructor(private code: Code, private quickopen: QuickOpen) { }
 
 	async showTerminal(): Promise<void> {
-		await this.code.dispatchKeybinding('ctrl+`');
+		await this.quickopen.runCommand('View: Toggle Integrated Terminal');
 		await this.code.waitForActiveElement(XTERM_TEXTAREA);
 		await this.code.waitForTerminalBuffer(XTERM_SELECTOR, lines => lines.some(line => line.length > 0));
 	}
