@@ -12,6 +12,6 @@ export abstract class Viewlet {
 	constructor(protected code: Code) { }
 
 	async waitForTitle(fn: (title: string) => boolean): Promise<void> {
-		await this.code.waitForTextContent('.monaco-workbench-container .part.sidebar > .title > .title-label > span', undefined, fn);
+		await this.code.waitForTextContent('.monaco-workbench .part.sidebar > .title > .title-label > h2', undefined, fn);
 	}
 }
