@@ -5,13 +5,22 @@
 
 const path = require('path');
 const Mocha = require('mocha');
+const minimist = require('minimist');
 
 const suite = 'Smoke Tests';
+
+const [, , ...args] = process.argv;
+const opts = minimist(args, {
+	string: [
+		'f'
+	]
+});
 
 const options = {
 	useColors: true,
 	timeout: 60000,
-	slow: 30000
+	slow: 30000,
+	grep: opts['f']
 };
 
 if (process.env.BUILD_ARTIFACTSTAGINGDIRECTORY) {
